requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy>=2.0.36
alembic==1.13.1
bcrypt==4.1.2
pydantic>=2.10.0
pydantic[email]>=2.10.0
python-dotenv==1.0.0